backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
pydantic
python-dotenv
pdfplumber
faiss-cpu
sentence-transformers
Pillow
openrouter
